requirements.txt
langchain
langgraph
langchain-openai
langchain-community
streamlit
chromadb
pypdf
openai
python-dotenv
sqlalchemy
fpdf
langchain-chroma
